# openapi/openapi.yaml
openapi: 3.1.0
info:
  description: External API endpoints for Artie
  title: Artie API
  version: v1.0.53
servers:
- url: https://api.artie.com
paths:
...
  /connectors/{uuid}:
    delete:
      description: Permanently removes a connector and its associated configuration.
        This stops any active replication pipelines that depend on the connector.
      parameters:
      - in: path
        name: uuid
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Delete a connector
      tags:
      - Connectors
      x-artie-mcp:
        destructiveHint: true
        idempotentHint: false
        openWorldHint: false
        readOnlyHint: false
    get:
      description: Retrieves the full configuration and status of a connector by its
        unique identifier, including connection settings and current replication state.
      parameters:
      - in: path
        name: uuid
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayloadsFullConnector'
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get a connector
      tags:
      - Connectors
      x-artie-mcp:
        destructiveHint: false
        idempotentHint: true
        openWorldHint: false
        readOnlyHint: true
    post:
      description: Updates the configuration of an existing source or destination
        connector. Modify connection settings, credentials, or replication parameters
        for a running pipeline.
      parameters:
      - in: path
        name: uuid
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayloadsConnectorPayload'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayloadsFullConnector'
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update a connector
      tags:
      - Connectors
      x-artie-mcp:
        destructiveHint: false
        idempotentHint: false
        openWorldHint: false
        readOnlyHint: false
...
components:
  schemas:
    EnumsConnectorSlug:
      enum:
      - api
      - bigquery
      - clickhouse
      - cockroach
      - databricks
      - documentdb
      - delta
      - dynamodb
      - gcs
      - iceberg
      - keyspaces
      - mongodb
      - motherduck
      - mssql
      - mysql
      - oracle
      - planetscale
      - postgresql
      - redis
      - redshift
      - s3
      - snowflake
      type: string
...
    PayloadsConnectorPayload:
      properties:
        dataPlaneName:
          type: string
        defaultDatabase:
          type:
          - "null"
          - string
        environmentUUID:
          format: uuid
          type:
          - "null"
          - string
        label:
          type: string
        privateLinkUUID:
          format: uuid
          type:
          - "null"
          - string
        sharedConfig:
          additionalProperties: {}
          type:
          - object
          - "null"
        snapshotPrivateLinkUUID:
          format: uuid
          type:
          - "null"
          - string
        sshTunnelUUID:
          format: uuid
          type:
          - "null"
          - string
        type:
          type: string
        uuid:
          type: string
      type: object
...
    PayloadsFullConnector:
      properties:
        companyUUID:
          format: uuid
          type: string
        connectionLabel:
          type: string
        createdAt:
          format: date-time
          type: string
        dataPlaneName:
          type:
          - "null"
          - string
        defaultDatabase:
          type: string
        environmentUUID:
          format: uuid
          type: string
        isValid:
          type: boolean
        label:
          type: string
        privateLinkUUID:
          format: uuid
          type:
          - "null"
          - string
        sharedConfig: {}
        snapshotPrivateLinkUUID:
          format: uuid
          type:
          - "null"
          - string
        sshTunnelUUID:
          format: uuid
          type:
          - "null"
          - string
        type:
          $ref: '#/components/schemas/EnumsConnectorSlug'
        updatedAt:
          format: date-time
          type: string
        uuid:
          format: uuid
          type: string
      required:
      - label
      - uuid
      - createdAt
      - updatedAt
      - type
      - companyUUID
      - environmentUUID
      - isValid
      type: object
...
  securitySchemes:
    ApiKey:
      description: API key passed as a Bearer token in the Authorization header.
      scheme: bearer
      type: http
security:
- ApiKey: []